refactor(sandbox): separate cleanup policy and evidence
- Move task recycling policy out of sandbox control.
- Distinguish pending evidence from terminal deletion authority.
- Share directory stages and journal advancement.
